Extract access token functions from notifier helper
Obtaining an access token from the OAuth2 token endpoint is a distinct operation from sending notifications.
This commit extracts jwt access token generation and supporting functions.
