+++ src/functions/notify/notifier.py
@@ -1,11 +1,9 @@
-import datetime
+import access_token
 import hashlib
-import jwt
 import logging
 import os
 import requests
 import routing_plans
-import time
 import uuid
 
 
@@ -13,7 +11,7 @@ def send_messages(data: dict) -> str:
     responses: list = []
     routing_plan_id: str = None
 
-    access_token: str = get_access_token()
+    token: str = access_token.get_token()
 
     if "routing_plan" in data:
         routing_plan_id = routing_plans.get_id(data.pop("routing_plan"))
@@ -23,16 +21,16 @@ def send_messages(data: dict) -> str:
             if "routing_plan" in message_data:
                 routing_plan_id = routing_plans.get_id(message_data.pop("routing_plan"))
 
-            response: str = send_message(access_token, routing_plan_id, message_data)
+            response: str = send_message(token, routing_plan_id, message_data)
             responses.append(response)
 
     return "\n".join(responses)
 
 
-def send_message(access_token, routing_plan_id, message_data) -> str:
+def send_message(token, routing_plan_id, message_data) -> str:
     body: str = message_body(routing_plan_id, message_data)
     correlation_id: str = message_data["correlation_id"]
-    response = requests.post(url(), json=body, headers=headers(access_token, correlation_id))
+    response = requests.post(url(), json=body, headers=headers(token, correlation_id))
 
     logging.info(f"Response from Notify API {url()}: {response.status_code}")
     if response.status_code == 201:
@@ -43,12 +41,12 @@ def send_message(access_token, routing_plan_id, message_data) -> str:
     return response.text
 
 
-def headers(access_token: str, correlation_id: str) -> dict:
+def headers(token: str, correlation_id: str) -> dict:
     return {
         "content-type": "application/vnd.api+json",
         "accept": "application/vnd.api+json",
         "x-correlation-id": correlation_id,
-        "authorization": "Bearer " + access_token,
+        "authorization": "Bearer " + token,
     }
 
 
@@ -89,82 +87,3 @@ def message_body(routing_plan_id, message_data) -> dict:
 def reference_uuid(val) -> str:
     str_val = str(val)
     return str(uuid.UUID(hashlib.md5(str_val.encode()).hexdigest()))
-
-
-def get_access_token() -> str:
-    if not os.getenv("OAUTH2_API_KEY"):
-        return "awaiting-token"
-
-    jwt: str = generate_auth_jwt()
-    headers: dict = {"Content-Type": "application/x-www-form-urlencoded"}
-
-    body = {
-        "grant_type": "client_credentials",
-        "client_assertion_type": "urn:ietf:params:oauth:client-assertion-type:jwt-bearer",
-        "client_assertion": jwt,
-    }
-
-    response = requests.post(
-        os.getenv("OAUTH2_TOKEN_URL"),
-        data=body,
-        headers=headers,
-    )
-    logging.info(f"Response from OAuth2 token provider: {response.status_code}")
-    response_json = response.json()
-
-    if response.status_code == 200:
-        access_token = response_json["access_token"]
-    else:
-        access_token = ""
-        logging.error("Failed to get access token")
-        logging.error(response_json)
-
-    return access_token
-
-
-def generate_auth_jwt() -> str:
-    algorithm: str = "RS512"
-    headers: dict = {
-        "alg": algorithm,
-        "typ": "JWT",
-        "kid": os.getenv("OAUTH2_API_KID")
-    }
-    api_key: str = os.getenv("OAUTH2_API_KEY")
-
-    payload: dict = {
-        "sub": api_key,
-        "iss": api_key,
-        "jti": str(uuid.uuid4()),
-        "aud": os.getenv("OAUTH2_TOKEN_URL"),
-        "exp": int(time.time()) + 300,  # 5mins in the future
-    }
-
-    private_key = get_private_key(os.getenv("PRIVATE_KEY_PATH", "private.pem"))
-
-    return generate_jwt(
-            algorithm, private_key, headers,
-            payload, expiry_minutes=5
-        )
-
-
-def generate_jwt(
-    algorithm: str,
-    private_key,
-    headers: dict,
-    payload: dict,
-    expiry_minutes: int = None,
-) -> str:
-    if expiry_minutes:
-        expiry_date = (
-            datetime.datetime.now(datetime.timezone.utc) +
-            datetime.timedelta(minutes=expiry_minutes)
-        )
-        payload["exp"] = expiry_date
-
-    return jwt.encode(payload, private_key, algorithm, headers)
-
-
-def get_private_key(private_key_path: str) -> str:
-    with open(private_key_path, "r", encoding="utf-8") as f:
-        private_key = f.read()
-        return private_key
